requirements.txt
Django==2.2.7
djangorestframework==3.10.3
python-memcached==1.59
